Release Version 1.0 complete - Ready for code freeze
This update integrates an update to the graphics engine that improves
performance, the ability to scale the scene via controller input, a
cursor that displays the color, size and shape of whatever the user will
brush, and an update to the UI that allows the user to interact with it
via a pointer beam from the right controller. This update includes
efforts by all three team members. Please check the readme file for
instructions for operating the program.